pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "A RAG application using Flask, Gemini, and ChromaDB."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core LangChain stack
    "langchain==0.2.5",
    "langchain-core==0.2.14",
    "langchain-community==0.2.4",
    "langchain-text-splitters==0.2.1",

    # Gemini integration (compatible pair)
    "langchain-google-genai==3.0.0",
    "google-generativeai==0.7.0",

    #For JinaAPI
    "requests==2.31.0",

    # Vector database
    "chromadb==0.4.24",

    # PDF parsing
    "PyMuPDF==1.24.7",
    "pypdf==4.2.0",

    # Utilities
    "numpy==1.26.4",

    # Flask backend
    "flask==3.0.3",
    "flask-cors==4.0.1",
    "gunicorn==22.0.0",
    "python-dotenv==1.0.1",

    # Logging and retries
    "tenacity==8.5.0",
    "packaging==23.2"
]
